Reverse the string passed through the function
This version of the code breaks down all the
steps necessary from converting a string given
to the function and reversing the order of the
characters present in the string.

Also, this verison of the code passes the first
test defined in the spec.js file.

Please note, I will be removing the comments/
pseudocode, and consolidating the code to make
it more readable.